+++ projectroles/rules.py
@@ -1,6 +1,9 @@
 import rules
 
-from .models import RoleAssignment, SODAR_CONSTANTS
+from django.conf import settings
+
+from projectroles.models import RoleAssignment, RemoteProject, SODAR_CONSTANTS
+
 
 # SODAR constants
 PROJECT_ROLE_OWNER = SODAR_CONSTANTS['PROJECT_ROLE_OWNER']
@@ -8,9 +11,11 @@ PROJECT_ROLE_DELEGATE = SODAR_CONSTANTS['PROJECT_ROLE_DELEGATE']
 PROJECT_ROLE_CONTRIBUTOR = SODAR_CONSTANTS['PROJECT_ROLE_CONTRIBUTOR']
 PROJECT_ROLE_GUEST = SODAR_CONSTANTS['PROJECT_ROLE_GUEST']
 PROJECT_TYPE_CATEGORY = SODAR_CONSTANTS['PROJECT_TYPE_CATEGORY']
+SITE_MODE_SOURCE = SODAR_CONSTANTS['SITE_MODE_SOURCE']
+SITE_MODE_TARGET = SODAR_CONSTANTS['SITE_MODE_TARGET']
 
 
-# Predicates -------------------------------------------------------------
+# Predicates -------------------------------------------------------------------
 
 
 @rules.predicate
@@ -77,65 +82,95 @@ def has_roles(user):
     return RoleAssignment.objects.filter(user=user).count() > 0
 
 
-# Rules ------------------------------------------------------------------
+@rules.predicate
+def is_modifiable_project(user, obj):
+    """Whether or not project metadata is modifiable"""
+    if settings.PROJECTROLES_SITE_MODE == SITE_MODE_TARGET:
+        try:
+            remote_project = RemoteProject.objects.get(
+                project=obj, site__mode=SITE_MODE_SOURCE)
+            return False
+
+        except RemoteProject.DoesNotExist:
+            pass
+
+    return True
+
+
+@rules.predicate
+def can_create_projects():
+    """Whether or not new projects can be generated on the site"""
+    if (settings.PROJECTROLES_SITE_MODE == SITE_MODE_TARGET and
+            not settings.PROJECTROLES_TARGET_CREATE):
+        return False
+
+    return True
+
+# Combined predicates ----------------------------------------------------------
+
+
+is_update_user = rules.is_superuser | is_project_owner | is_project_delegate
+
+
+# Rules ------------------------------------------------------------------------
 
 
 # Rules should not be needed, use permissions for user rights
 
 
-# Permissions ------------------------------------------------------------
+# Permissions ------------------------------------------------------------------
 
 
 # Allow viewing project/category details
 rules.add_perm(
     'projectroles.view_project',
-    rules.is_superuser | has_project_role | has_category_child_role)
+    has_project_role | has_category_child_role)
 
 # Allow project updating
 rules.add_perm(
     'projectroles.update_project',
-    rules.is_superuser | is_project_owner | is_project_delegate)
+    is_update_user & is_modifiable_project)
 
 # Allow creation of projects
 rules.add_perm(
     'projectroles.create_project',
-    rules.is_superuser | is_project_owner)
+    is_project_owner & can_create_projects)
 
 # Allow updating project settings
 rules.add_perm(
     'projectroles.update_project_settings',
-    rules.is_superuser | is_project_owner | is_project_delegate)
+    is_update_user & is_modifiable_project)
 
 # Allow viewing project roles
 rules.add_perm(
     'projectroles.view_project_roles',
-    rules.is_superuser | is_project_owner | is_project_delegate |
+    is_project_owner | is_project_delegate |
     is_project_contributor | is_project_guest)
 
 # Allow updating project owner
 rules.add_perm(
     'projectroles.update_project_owner',
-    rules.is_superuser | is_project_owner)
+    is_project_owner & is_modifiable_project)
 
 # Allow updating project delegate
 rules.add_perm(
     'projectroles.update_project_delegate',
-    rules.is_superuser | is_project_owner)
+    is_project_owner & is_modifiable_project)
 
 # Allow updating project members
 rules.add_perm(
     'projectroles.update_project_members',
-    rules.is_superuser | is_project_owner | is_project_delegate)
+    is_update_user & is_modifiable_project)
 
 # Allow inviting users to project via email
 rules.add_perm(
     'projectroles.invite_users',
-    rules.is_superuser | is_project_owner | is_project_delegate)
+    is_update_user & is_modifiable_project)
 
 # Allow importing roles from another project
 rules.add_perm(
     'projectroles.import_roles',
-    rules.is_superuser | is_project_owner)
+    is_project_owner & is_modifiable_project)
 
 # Allow updating remtote sites and remote project access
 rules.add_perm(